Pause Menu
Fixed a bug with wall collision - continuously hugging the wall no
longer slows down the character.

Added a pause menu to level_design. This is bound to the 'escape' key.
Pressing escape again or clicking resume resumes the game. The main menu
button takes us to the main screen. Quit will exit the application,
which only works in an actual build, does nothing in editor.